+++ graphql-spqr-spring-boot-autoconfigure/src/main/java/io/leangen/graphql/spqr/spring/web/dto/TransportType.java
@@ -1,0 +1,9 @@
+package io.leangen.graphql.spqr.spring.web.dto;
+
+public enum TransportType {
+    HTTP, HTTP_EVENT_STREAM, WEBSOCKET;
+
+    public boolean isEventStream() {
+        return this != HTTP;
+    }
+}