+++ graphql-spqr-spring-boot-autoconfigure/src/main/java/io/leangen/graphql/spqr/spring/web/servlet/websocket/GraphQLWebSocketExecutor.java
@@ -3,12 +3,12 @@ package io.leangen.graphql.spqr.spring.web.servlet.websocket;
 import graphql.ExecutionResult;
 import graphql.GraphQL;
 import io.leangen.graphql.spqr.spring.web.GraphQLExecutor;
-import io.leangen.graphql.spqr.spring.web.dto.GraphQLRequest;
+import io.leangen.graphql.spqr.spring.web.dto.ExecutorParams;
 import org.springframework.web.socket.WebSocketSession;
 
 @FunctionalInterface
 public interface GraphQLWebSocketExecutor extends GraphQLExecutor<WebSocketSession> {
 
     @Override
-    ExecutionResult execute(GraphQL graphQL, GraphQLRequest graphQLRequest, WebSocketSession request);
+    ExecutionResult execute(GraphQL graphQL, ExecutorParams<WebSocketSession> params);
 }