+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support subscriptions over SSE [#97](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/97)
+
+### Changed
+- Rewrite the WebSocket-based subscriptions implementation [#98](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/98)
+
 ## [0.0.5] - 2021-02-09
 ### Added
 - Support for Spring Data paging 🥳, with ot without Relay compatibility (thanks [@vjroby](https://github.com/vjroby)) [#2](https://github.com/leangen/graphql-spqr-spring-boot-starter/issues/2)